updated tournament and member

+++ src/main/java/com/keyin/tournament/TournamentService.java
@@ -1,14 +1,11 @@
 package com.keyin.tournament;
 
-
 import com.keyin.member.Member;
 import com.keyin.member.MemberRepository;
-import com.keyin.tournament.TournamentDTO;
-import com.keyin.tournament.Tournament;
-import com.keyin.tournament.TournamentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -27,8 +24,11 @@ public class TournamentService {
         tournament.setStartDate(tournamentDTO.getStartDate());
         tournament.setEndDate(tournamentDTO.getEndDate());
         tournament.setLocation(tournamentDTO.getLocation());
-        tournament.setEntryFee(tournamentDTO.getEntryFee());
-        tournament.setCashPrize(tournamentDTO.getCashPrize());
+
+        // Convert entry fee and cash prize to BigDecimal if they are double in DTO
+        tournament.setEntryFee(BigDecimal.valueOf(tournamentDTO.getEntryFee()));
+        tournament.setCashPrize(BigDecimal.valueOf(tournamentDTO.getCashPrize()));
+
 
         Set<Member> members = new HashSet<>(memberRepository.findAllById(tournamentDTO.getMemberIds()));
         tournament.setParticipatingMembers(members);
@@ -40,4 +40,3 @@ public class TournamentService {
         return tournamentRepository.findAll();
     }
 }
-
